motion and insert commands.
